AnularSeed Project - Template Store Pal
AngularJS site to list the templates

Using AngularSee and Bootstrap Boilerplate to develop the site.

Site is so simple, on the first page we will list all the template
themes, on click of each one will take you to the details page where we
will show case variations of template themes with images. Also we have
PayPal Sandbox button to purchase the theme.